BugFix: Update the return type of the closest function
As the MDN spec for the .closest function states, the return type is 'The closest ancestor Element or itself, which matches the selectors. If there are no such element, null' this means it should be an actual Element type and not a node. This solves a wider issue with TS that prevents users from chaining querySelectors and calls to closest.